data-scout-agent/pyproject.toml
[project]
name = "scout-agent"
version = "0.1.0"
description = "A multi-agent system for discovering, curating, and auditing training data."
authors = [{ name = "Jules", email = "jules@example.com" }]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "langgraph",
    "langgraph-checkpoint-sqlite",
    "langchain-openai",
    "attachments",
    "pydantic",
    "beautifulsoup4",
    "litellm",
    "langsmith", # Corrected from langsmith-sdk to langsmith for PyPI
    "requests",
    "beautifulsoup4",
    "html5lib",
    "python-dotenv"
]

[project.optional-dependencies]
crawler = ["libcrawler"]

[project.scripts]
scout-agent = "scout_agent.main:run"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
